+++ R/generics.R
@@ -33,10 +33,10 @@ setGeneric(
 )
 
 
-# Methods and documentations found in methods-spatShift.R
+
 setGeneric("spatShift", function(x, ...) standardGeneric("spatShift"))
-# Methods and documentations found in methods-shear.R
 setGeneric("shear", function(x, ...) standardGeneric("shear"))
+setGeneric("affine", function(x, ...) standardGeneric("affine"))
 
 # Methods and documentations found in methods-overlaps.R
 setGeneric("overlaps", function(x, ...) standardGeneric("overlaps"))