+++ NEWS.md
@@ -16,6 +16,10 @@
 - `installGiottoEnvironment()`, `removeGiottoEnvironment()` now have `conda` param for setting path to conda executable and `envname` param for specifying environment by name
 - `installGiottoEnvironment()` now has `confirm` param for skipping path input checks
 
+## new
+- `affine()` for `giottoPolygon`, `giottoPoints`, `spatLocsObj`
+- `decomp_affine()` for decomposing affine matrix to simple transforms
+
 
 # GiottoClass 0.3.1 (2024/05/21)